Convert ticks to 64-bit, added nanosecond precision to the API
Fixes https://github.com/libsdl-org/SDL/issues/5512
Fixes https://github.com/libsdl-org/SDL/issues/6731

+++ src/events/SDL_mouse_c.h
@@ -38,7 +38,7 @@ typedef struct
 typedef struct
 {
     int last_x, last_y;
-    Uint32 last_timestamp;
+    Uint64 last_timestamp;
     Uint8 click_count;
 } SDL_MouseClickState;
